change particle vector to particle list
the access to a special entry in a vector is constant but deleting or
inserting at a special point costs linear time. We don't need any
access to a special particle, we iterate every time through all the
vector. Therefore, the container list is much better. It allows
constant deleting and inserting